Rebase w/ new placement of parsing files
