+++ src/ert/_c_wrappers/job_queue/workflow_job.py
@@ -10,7 +10,6 @@ from ert._c_wrappers.job_queue import ErtScript, ExternalErtScript, FunctionErtS
 from ert._clib.job_kw import type_from_kw
 from ert.parsing import ConfigValidationError
 
-from ...parsing import ConfigValidationError
 from .ert_plugin import ErtPlugin
 
 if TYPE_CHECKING:
