+++ src/ert/parsing/__init__.py
@@ -1,5 +1,13 @@
 from .config_errors import ConfigValidationError, ConfigWarning
 from .lark_parser import parse as lark_parse
-from .lark_parser_types import ErrorInfo
+from .lark_parser_error_info import ErrorInfo
+from .lark_parser_types import MaybeWithKeywordToken, MaybeWithToken
 
-__all__ = ["lark_parse", "ConfigWarning", "ConfigValidationError", "ErrorInfo"]
+__all__ = [
+    "lark_parse",
+    "ConfigWarning",
+    "ConfigValidationError",
+    "ErrorInfo",
+    "MaybeWithToken",
+    "MaybeWithKeywordToken",
+]
