+++ src/ert/parsing/config_errors.py
@@ -1,4 +1,6 @@
-from typing import Optional
+from typing import List, Optional
+
+from .lark_parser_error_info import ErrorInfo
 
 
 class ConfigWarning(UserWarning):
@@ -17,3 +19,14 @@ class ConfigValidationError(ValueError):
             if self.config_file
             else f"{self.errors}"
         )
+
+    @classmethod
+    def raise_from_collected(cls, collected_errors: List[ErrorInfo]):
+        if len(collected_errors) > 0:
+            combined_str = ";".join([x.message for x in collected_errors])
+            first_filename = next(
+                x.filename for x in collected_errors if x.filename is not None
+            )
+            raise ConfigValidationError(
+                errors=combined_str, config_file=first_filename or None
+            )
